requirements.txt
openai
langchain
chromadb
tiktoken
beautifulsoup4
requests
schedule
python-dotenv
